feat(i18n): move flattenKeys/unflattenKeys from i18n into core module — remove old i18n folder
