+++ packages/i18nify-js/src/modules/core/__tests__/flattenKeys.test.ts
