+++ packages/i18nify-js/src/modules/core/constants.ts
