+++ packages/i18nify-js/src/modules/core/flattenKeys.ts
