+++ packages/i18nify-js/src/modules/core/index.ts
@@ -1,3 +1,7 @@
 export { default as getState } from './getState';
 export { default as setState } from './setState';
 export { default as resetState } from './resetState';
+export { default as flattenKeys } from './flattenKeys';
+export { default as unflattenKeys } from './unflattenKeys';
+export type { JsonObject, FlattenOptions, UnflattenOptions } from './types';
+export { DEFAULT_DELIMITER } from './constants';
