+++ packages/i18nify-js/src/modules/core/types.ts
