--- packages/i18nify-js/src/modules/i18n/index.ts
@@ -1,4 +1,0 @@
-export { default as flattenKeys } from './flattenKeys';
-export { default as unflattenKeys } from './unflattenKeys';
-export type { FlattenOptions, UnflattenOptions, JsonObject } from './types';
-export { DEFAULT_DELIMITER } from './constants';
